requirements.txt
playwright==1.49.1
python-dotenv==1.0.1
python-telegram-bot==21.9
